backend/pyproject.toml
[tool.poetry]
name = "docintell-backend"
version = "0.1.0"
description = "AI-powered document intelligence platform"
authors = ["Developer"]
packages = [{include = "app"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
redis = "^5.0.1"
celery = "^5.3.4"
pypdf2 = "^3.0.1"
pytesseract = "^0.3.10"
pillow = "^10.2.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
langchain-community = "^0.0.10"
chromadb = "^0.4.22"
openai = "^1.8.0"
tiktoken = "^0.5.2"
numpy = "^1.26.3"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
prometheus-client = "^0.19.0"
structlog = "^24.1.0"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
factory-boy = "^3.3.0"
faker = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "C90", "UP", "B", "A", "C4", "DJ", "EM", "ISC", "PIE", "PL", "PT", "Q", "RET", "SIM", "TID", "TCH", "ARG", "DTZ", "ERA", "PD", "PGH", "RUF"]
ignore = ["E501", "B008", "RUF009"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"